Keybindings: move input_service:get(special_key) calls out of the loop.
The function is called whenever a key is pressed, therefore we only need to check for ctrl, alt and shift once. Each of them was called twice, nested inside two loops.

The same thing can be found in Custom Menus, the same block of three lines and six conditionals. Good target for refactoring.