Allow special characters in mod settings search
Previously, users could not type % in the search bar. Typing [, ], (, or ) would crash, and . + - * ? ^ $ were not escaped so you could not search for those characters, but instead they would perform their typical Lua pattern functions which would be confusing to the user.

Simply escape all of these characters so that the user can search for them if they want, and no more crashes!