Allow multiple keybind presses at once
It looks like VMF keybinds were originally designed to only support a single key at a time. I wasn't able to figure out why this was, and no one I asked seem to think that it's important for it to work that way. Personally, I also noticed some annoying behavior, which is that other keybinds would not work for ~0.25 seconds after releasing a different keybind.

I tested this with both the "held" and "pressed" varieties of keybinds and did not experience any issues. I can now press keybinds while holding another down though.